Refactor controllers/model to User model
Previously our user model used two different tables(participants and researchers) to
create the two roles. This refactor moves participants to the user class where
researcher will join them in a future branch. A separate user controller was
nested under participants, and attributes were sorted into User attributes
which will be shared with researchers, and a ParticipantCredential class to
store tokens.

Tests were written for these models and for the user service and authentication
service.

Future refactors that will have to build off this are:
* Adding a password in a separate view/path to store on the participant-user
* Adding a logout
* Adding a login that doesn't hit the api
* Adding a token refresh method