Add logout link
Strangely log out works, but not in a test. I believe it's because I am only
mocking that the user be logged in since I am currently just setting the session
hash in one controller for one type of user. Will come back to this test once
I have a login that exists outside of the authentication process.