Add researcher-user creation
Primary addition was a standard CRUD-style creation of a researcher-user with a
form that also pulls in their researcher credentials/profile. Landing page now
directs a visitor to choose what kind of account they want to create, or if they
want to login.

Bcrypt was implemented for passwords and participant-users have a default password
for now because the entry doesn't save without a password.