Adding better dashboard paths
Add study/snp/snp_value creation

Large commit that creates the controller, view(form), and models for the pieces
associated with creating a study. The model uses nested attributes on the study
controller. Currently a researcher only sees studies in their own dashboard but
a refactor could bring them out to be public.

It was also necessary to refactor participant/researcher dashboards as a part of
this. It seemed better to render separate views and use the same controller
rather than have a lot of code in the view.

Closes #16