Add service, method, view for Participant matches
Created a method for matching participants with studies where they both have the
same snp value, i.e., they have the same base pairs at a given location.

Required adding a new endpoint and service for 23 and Me. Tested the service that
was created for this. Added a feature test for a user's matches. Created a view
for the study which is currently empty, so needs a feature test for that.

Code could also be refactored to use AR queries.

Added activerecord-import to handle bulk importing snp locations to the database. Changed rake
task to reflect this.