Add api route for snp verification
Added api route for verification of a snp's location. When a researcher creates
a new study, the form automatically verifies if a location is valid and provides
visual confirmation of such (or an "X" if it is invalid).

The endpoint is tested at the service level but a feature test could be used to
verify the ajax confirmation.