Added cron job to update access token once a day
Using a gem that adds cron job functionality to Sidekiq, added a job to update
access tokens for all participants.

Initially it seemed to be proccing jobs more than necessary, but after killing
the sidekiq server it's more calm on rebooting it. Perhap the debugging had
generated a lot of queued jobs.