Add styling, fix affected tests and study creation bug
Add basic styling. In changing some of the language, this affected a couple of
tests.

Additionally I realized that study creation was a little buggy due to how
first or create works (or doesn't work) with nested attributes. To compensate,
the SnpValue class now has a method to find the location before validation.