Refactor participant/user controller with new class
Refactored user creation out of the auth path for the participants/user
controller with a new UserCreator class. While there currently is a method
in the User class to generate a new user, this new class also takes on the
functionality of contacting the Auth service and adding a participant credential.