Fix bug with duplicate snp_values
Due to the way I was nesting associations in a form and then
handling new study creation, users would never get matched up with
subsequent studies that had the same locations & base pairs as a
previous study. Added a new service to break apart study creation.
Could be an avenue for future cleanup.